Redo field updates for faster compile
